Fix varlink GetVersion()
Fixed two small problems with the varlink GetVersion() call.  The first
was that an error was being eaten and not returned to users when a
string to int conversion was being done.  The second was that the
variable given to the string to int conversion was not something that
could be converted into an integer (i.e. 2.1.0).  Adjusted the
APIVersion to be only representative of the major number (i.e. 2).

Fixes bugzilla: #1925928

Signed-off-by: baude <bbaude@redhat.com>