[CI:DOCS] Restore rootless-cni-infra contrib info
PR #9423 removed the contrib/rootless-cni-infra directory and
it's files and should not have.  The information in that tree
is required for Podman v3.0 and earlier users, and primarily
those in RHEL 8.4.0 and 8.3.1.

This PR puts those files back and adds a note to the README.md
noting the changes are not needed for versions after Podman v3.0.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>